Clean up and finalize add encomapss client extension AddEncompassHttpResponseHeaderLoggingHandler
