--- src/EncompassApi/Configuration/HttpResponseHeaderLogOptions.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace EncompassApi.Configuration
-{
-    public class HttpResponseHeaderLogOptions
-    {
-        public IEnumerable<string> Headers { get; set; }
-    }
-}
